Decouple functions, second request for show after save
When saving, the resulting edit buttons do not reflect the actual
range. That is why we simply ask the server again for the content. In
addition we need to unbind our loadEditDialog handler when super-evily
reloading the jQuery or we end up with two handlers on the same button
after closing without saving.